Refactor implicit_tpm fixture
